backend/SalutFinance-main/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
executing==2.0.1
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
icecream==2.1.3
idna==3.6
orjson==3.9.12
passlib==1.7.4
pyasn1==0.5.1
pydantic==2.6.3
pydantic-settings==2.1.0
pydantic_core==2.16.3
Pygments==2.17.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1